+++ common/src/main/java/com/kwwsyk/endinv/common/client/gui/page/DisplayPage.java
@@ -231,7 +231,7 @@ public abstract class DisplayPage{
 
     public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTick){
         renderPage(graphics);
-        if(framework.notHoveringOnSortBox()){
+        if(framework.hoveringOnPage()){
             renderHovering(graphics, mouseX, mouseY, partialTick);
         }
     }