+++ common/src/main/java/com/kwwsyk/endinv/common/client/gui/page/ItemPage.java
@@ -6,7 +6,6 @@ import com.kwwsyk.endinv.common.client.gui.EndlessInventoryScreen;
 import com.kwwsyk.endinv.common.client.option.CachedConfig;
 import com.kwwsyk.endinv.common.menu.page.PageType;
 import com.kwwsyk.endinv.common.menu.page.pageManager.PageMetaDataManager;
-import com.kwwsyk.endinv.common.network.payloads.PageData;
 import com.kwwsyk.endinv.common.network.payloads.toServer.CreativeItemModPayload;
 import com.kwwsyk.endinv.common.network.payloads.toServer.ItemClickPayload;
 import com.kwwsyk.endinv.common.network.payloads.toServer.ItemPageContext;
@@ -105,8 +104,7 @@ public abstract class ItemPage extends DisplayPage {
      * For ItemPage and ItemDisplay: also used to request contents as sending {@link ItemPageContext} has such side effect.
      */
     public void sendChangesToServer() {
-        PageData layout = meta.getPageData();
-        CachedConfig.updateLayout(layout);
+        var layout = CachedConfig.currentLayout();
         getPacketDistributor().sendToServer(new ItemPageContext(startIndex, length, layout));//send this packet will receive a content packet as callback
     }