+++ common/src/main/java/com/kwwsyk/endinv/common/client/gui/widget/SortTypeSwitchBox.java
@@ -1,8 +1,9 @@
 package com.kwwsyk.endinv.common.client.gui.widget;
 
-import com.kwwsyk.endinv.common.client.gui.SortTypeSwitcher;
+import com.kwwsyk.endinv.common.client.gui.ScreenFramework;
 import com.kwwsyk.endinv.common.client.gui.bg.ScreenRectangleWidgetParam;
 import com.kwwsyk.endinv.common.client.gui.page.manager.PageManager;
+import com.kwwsyk.endinv.common.client.option.CachedConfig;
 import com.kwwsyk.endinv.common.util.SortType;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
@@ -14,25 +15,21 @@ import org.jetbrains.annotations.NotNull;
 
 public class SortTypeSwitchBox extends AbstractWidget {
 
-    public SortTypeSwitcher screen;
+    public ScreenFramework screen;
     private final PageManager pageManager;
-    private int x;
-    private int y;
     private final int singleBoxHeight;
     private boolean isOpen;
 
-    
-    public SortTypeSwitchBox(SortTypeSwitcher screen, PageManager pageManager, int x, int y, int width, int height){
+
+    public SortTypeSwitchBox(ScreenFramework screen, PageManager pageManager, int x, int y, int width, int height){
         super(x,y,width,height, Component.empty());
         this.screen = screen;
         this.pageManager = pageManager;
         this.visible = pageManager.getDisplayingPageType() != null;
-        this.x = x;
-        this.y = y;
         this.singleBoxHeight = height;
     }
 
-    public SortTypeSwitchBox(SortTypeSwitcher screen, PageManager pageManager, ScreenRectangleWidgetParam sortTypeSwitchBoxParam){
+    public SortTypeSwitchBox(ScreenFramework screen, PageManager pageManager, ScreenRectangleWidgetParam sortTypeSwitchBoxParam){
         this(screen,
                 pageManager,
                 sortTypeSwitchBoxParam.XPos(),
@@ -55,7 +52,7 @@ public class SortTypeSwitchBox extends AbstractWidget {
         if(!isOpen){
             setOpen(true);
         }else {
-            int y1 = y+singleBoxHeight;
+            int y1 = getY()+singleBoxHeight;
             for(SortType type : SortType.values()){
                 if(isHoveringOnSingleBox((int)mouseY,y1)){
                     screen.switchSortTypeTo(type);
@@ -76,27 +73,26 @@ public class SortTypeSwitchBox extends AbstractWidget {
 
     @Override
     protected void renderWidget(@NotNull GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
-        if(isHovered) screen.setHoveringOnSortBox(true);
         guiGraphics.pose().pushPose();
         guiGraphics.pose().translate(0.0F,0.0F,500.0F);
-        guiGraphics.fill(x,y,x+width,y+singleBoxHeight,0xff888888);
-        guiGraphics.fill(x+1,y+1,x+width-1,y+singleBoxHeight-1,0xff000000);
-        if(isHoveringOnSingleBox(mouseY,y))
-            guiGraphics.fillGradient(RenderType.guiOverlay(),x,y,x+width,y+singleBoxHeight,0x80ffffff,0x80ffffff,0);
-        SortType sortType = pageManager.sortType();
+        guiGraphics.fill(getX(),getY(),getX()+width,getY()+singleBoxHeight,0xff888888);
+        guiGraphics.fill(getX()+1, getY() +1,getX()+width-1, getY() +singleBoxHeight-1,0xff000000);
+        if(isHoveringOnSingleBox(mouseY, getY()))
+            guiGraphics.fillGradient(RenderType.guiOverlay(),getX(), getY(),getX()+width, getY() +singleBoxHeight,0x80ffffff,0x80ffffff,0);
+        SortType sortType = CachedConfig.sortType();
         String s = sortType.toString();
-        guiGraphics.drawString(Minecraft.getInstance().font, s,x+2,y+2,0xffffffff);
+        guiGraphics.drawString(Minecraft.getInstance().font, s,getX()+2, getY() +2,0xffffffff);
         if(isOpen){
-            int y1 = y+singleBoxHeight;
+            int y1 = getY() +singleBoxHeight;
             for(SortType type : SortType.values()){
-                guiGraphics.fill(RenderType.gui(),x,y1,x+width,y1+singleBoxHeight,0,0xff888888);
-                guiGraphics.fill(RenderType.gui(),x+1,y1+1,x+width-1,y1+singleBoxHeight-1,0,0xff000000);
+                guiGraphics.fill(RenderType.gui(),getX(),y1,getX()+width,y1+singleBoxHeight,0,0xff888888);
+                guiGraphics.fill(RenderType.gui(),getX()+1,y1+1,getX()+width-1,y1+singleBoxHeight-1,0,0xff000000);
                 if(isHoveringOnSingleBox(mouseY,y1)) {
-                    guiGraphics.fillGradient(RenderType.guiOverlay(), x, y1, x + width, y1 + singleBoxHeight, 0x80ffffff, 0x80ffffff, 0);
+                    guiGraphics.fillGradient(RenderType.guiOverlay(), getX(), y1, getX() + width, y1 + singleBoxHeight, 0x80ffffff, 0x80ffffff, 0);
                     guiGraphics.renderTooltip(Minecraft.getInstance().font,Component.translatable(type.translationKey),mouseX,mouseY);
                 }
                 s = type.toString();
-                guiGraphics.drawString(Minecraft.getInstance().font, s,x+2,y1+2,0xffffffff);
+                guiGraphics.drawString(Minecraft.getInstance().font, s,getX()+2,y1+2,0xffffffff);
                 y1+=singleBoxHeight;
             }
         }